+++ src/services/index.ts
@@ -41,3 +41,5 @@ export * from "./blog.service";
 export * from "./seller.service";
 export * from "./promotions.service";
 export * from "./coupon.service";
+export * from "./demo.service";
+export * from "./realtime-token.service";